allow using custom artifact downloader
