+++ stub-runner/stub-runner/src/main/groovy/io/codearte/accurest/stubrunner/StubRunnerFactory.groovy
@@ -3,25 +3,22 @@ package io.codearte.accurest.stubrunner
 import groovy.transform.CompileStatic
 import groovy.transform.PackageScope
 import groovy.util.logging.Slf4j
+
 /**
  * Factory of StubRunners. Basing on the options and passed collaborators
  * downloads the stubs and returns a list of corresponding stub runners.
  */
 @Slf4j
 @CompileStatic
-@PackageScope
 class StubRunnerFactory {
 
 	private final StubRunnerOptions stubRunnerOptions
 	private final Collection<StubConfiguration> collaborators
 	private final StubDownloader stubDownloader
 
-	StubRunnerFactory(StubRunnerOptions stubRunnerOptions, Collection<StubConfiguration> collaborators) {
-		this(stubRunnerOptions, collaborators, new StubDownloader())
-	}
-
-	protected StubRunnerFactory(StubRunnerOptions stubRunnerOptions, Collection<StubConfiguration> collaborators,
-								StubDownloader stubDownloader) {
+	StubRunnerFactory(StubRunnerOptions stubRunnerOptions,
+	                  Collection<StubConfiguration> collaborators,
+	                            StubDownloader stubDownloader) {
 		this.stubRunnerOptions = stubRunnerOptions
 		this.collaborators = collaborators
 		this.stubDownloader = stubDownloader
